demodir: unset direnv env so direnv does not reload

+++ flake.nix
@@ -60,24 +60,12 @@
           toDemoShell = version: contrast-release:
             lib.nameValuePair "demo-${version}" (pkgs.mkShell {
               packages = [ contrast-release ];
-
               shellHook = ''
                 cd "$(mktemp -d)"
-
-                if [[ -e ${contrast-release}/runtime.yml ]]
-                then
-                  cp ${contrast-release}/runtime.yml .
-                fi
-
-                if [[ -e ${contrast-release}/coordinator.yml ]]
-                then
-                  cp ${contrast-release}/coordinator.yml .
-                fi
-                
-                if [[ -d ${contrast-release}/deployment ]]
-                then
-                  cp -r ${contrast-release}/deployment .
-                fi
+                [[ -e ${contrast-release}/runtime.yml ]] && cp ${contrast-release}/runtime.yml .
+                [[ -e ${contrast-release}/coordinator.yml ]] && cp ${contrast-release}/coordinator.yml .
+                [[ -d ${contrast-release}/deployment ]] && cp -r ${contrast-release}/deployment .
+                export DO_NOT_TRACK=1
               '';
             });
         in
